requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.1
Django==5.0.2
django-coverage==1.2.4
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pluggy==1.4.0
prompt-toolkit==3.0.43
pytest==8.0.1
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
